// client/src/types/spotify.ts
export type WebPlaybackStatuses = "ready" | "not_ready";
export type WebPlaybackStates = "player_state_changed";
export type WebPlaybackErrors =
  | "initialization_error"
  | "authentication_error"
  | "account_error"
  | "playback_error";

export interface IWebPlaybackError {
  message: WebPlaybackErrors;
}

export interface IWebPlaybackPlayer {
  _options: {
    getOAuthToken: () => () => void;
    name: string;
    id: string;
    volume: number;
  };
  addListener: {
    (
      event: WebPlaybackErrors,
      callback: (d: IWebPlaybackError) => void
    ): boolean;
    (
      event: WebPlaybackStates,
      callback: (d: IWebPlaybackState | null) => void
    ): boolean;
    (
      event: WebPlaybackStatuses,
      callback: (d: IWebPlaybackReady) => void
    ): boolean;
  };
  connect: () => Promise<boolean>;
  disconnect: () => void;
  getCurrentState: () => Promise<IWebPlaybackState | null>;
  getVolume: () => Promise<number>;
  nextTrack: () => Promise<void>;
  pause: () => Promise<void>;
  previousTrack: () => Promise<void>;
  removeListener: (
    event: WebPlaybackErrors | WebPlaybackStates | WebPlaybackStatuses,
    callback?: () => void
  ) => boolean;
  resume: () => Promise<void>;
  seek: (positionMS: number) => Promise<void>;
  setName: (n: string) => Promise<void>;
  setVolume: (n: number) => Promise<void>;
  togglePlay: () => Promise<void>;
}

export interface IWebPlaybackReady {
  device_id: string;
}

export interface IWebPlaybackState {
  context: {
    uri: null;
    metadata: object;
  };
  bitrate: number;
  position: number;
  duration: number;
  paused: boolean;
  shuffle: boolean;
  repeat_mode: number;
  track_window: {
    current_track: IWebPlaybackTrack;
    next_tracks: IWebPlaybackTrack[];
    previous_tracks: IWebPlaybackTrack[];
  };
  timestamp: number;
  restrictions: {
    disallow_resuming_reasons: [];
    disallow_skipping_prev_reasons: [];
  };
  disallows: {
    resuming: boolean;
    skipping_prev: boolean;
  };
}

export interface IWebPlaybackAlbum {
  uri: string;
  name: string;
  images: IWebPlaybackImage[];
}

export interface IWebPlaybackArtist {
  name: string;
  uri: string;
}

export interface IWebPlaybackImage {
  height: number;
  url: string;
  width: number;
}

export interface IWebPlaybackTrack {
  id: string;
  uri: string;
  type: string;
  linked_from_uri: null | string;
  linked_from: {
    uri: null | string;
    id: null | string;
  };
  media_type: string;
  name: string;
  duration_ms: number;
  artists: IWebPlaybackArtist[];
  album: IWebPlaybackAlbum;
  is_playable: boolean;
}